# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Enterprise AI Agent",
    "description": "Enterprise LLM Application Platform — 企业级 Agent 与知识助手 REST API。\n\n- **Swagger UI**: `/docs`\n- **ReDoc**: `/redoc`\n- **OpenAPI JSON**: `/openapi.json`\n\n详细说明见仓库 `docs/API.md`。",
    "contact": {
      "name": "Enterprise AI Platform",
      "url": "https://github.com/your-org/enterprise-ai-agent"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/workflow": {
      "get": {
        "tags": [
          "Agent Studio"
        ],
        "summary": "Get Workflow",
        "operationId": "get_workflow_workflow_get",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workflow Id"
            }
          },
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowViewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/trace": {
      "get": {
        "tags": [
          "Agent Studio"
        ],
        "summary": "Get Trace",
        "operationId": "get_trace_trace_get",
        "parameters": [
          {
            "name": "trace_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trace Id"
            }
          },
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceViewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/prompt": {
      "get": {
        "tags": [
          "Agent Studio"
        ],
        "summary": "Get Prompt",
        "operationId": "get_prompt_prompt_get",
        "parameters": [
          {
            "name": "prompt_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Id"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptViewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Agent Studio"
        ],
        "summary": "Update Prompt",
        "operationId": "update_prompt_prompt_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePromptRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptViewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/memory": {
      "get": {
        "tags": [
          "Agent Studio"
        ],
        "summary": "Get Memory",
        "operationId": "get_memory_memory_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryViewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/agent/status": {
      "get": {
        "tags": [
          "Agent Studio"
        ],
        "summary": "Get Agent Status",
        "operationId": "get_agent_status_agent_status_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workflow/rerun": {
      "post": {
        "tags": [
          "Agent Studio"
        ],
        "summary": "Rerun Workflow Step",
        "operationId": "rerun_workflow_step_workflow_rerun_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RerunStepRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowViewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/gpu/allocate": {
      "post": {
        "tags": [
          "GPU Resource Manager"
        ],
        "summary": "Gpu Allocate",
        "operationId": "gpu_allocate_api_v1_gpu_allocate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GpuAllocationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GpuAllocationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/gpu/allocate/schedule": {
      "post": {
        "tags": [
          "GPU Resource Manager"
        ],
        "summary": "Gpu Allocate Scheduled",
        "operationId": "gpu_allocate_scheduled_api_v1_gpu_allocate_schedule_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GpuAllocationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GpuAllocationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/gpu/release/{allocation_id}": {
      "post": {
        "tags": [
          "GPU Resource Manager"
        ],
        "summary": "Gpu Release",
        "operationId": "gpu_release_api_v1_gpu_release__allocation_id__post",
        "parameters": [
          {
            "name": "allocation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Allocation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Gpu Release Api V1 Gpu Release  Allocation Id  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/service/providers/register": {
      "post": {
        "tags": [
          "Service Discovery"
        ],
        "summary": "Register Provider",
        "operationId": "register_provider_api_v1_service_providers_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderNode"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/service/providers/{node_id}": {
      "delete": {
        "tags": [
          "Service Discovery"
        ],
        "summary": "Unregister Provider",
        "operationId": "unregister_provider_api_v1_service_providers__node_id__delete",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Unregister Provider Api V1 Service Providers  Node Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/service/heartbeat": {
      "post": {
        "tags": [
          "Service Discovery"
        ],
        "summary": "Provider Heartbeat",
        "operationId": "provider_heartbeat_api_v1_service_heartbeat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderNode"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentStatusResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "workflow_id": {
            "type": "string",
            "title": "Workflow Id",
            "default": ""
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": ""
          },
          "agents": {
            "additionalProperties": true,
            "type": "object",
            "title": "Agents"
          },
          "pending_node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Node Id"
          }
        },
        "type": "object",
        "required": [
          "run_id"
        ],
        "title": "AgentStatusResponse"
      },
...
      "GpuAllocationRequest": {
        "properties": {
          "memory_mib": {
            "type": "number",
            "minimum": 1.0,
            "title": "Memory Mib",
            "default": 4096
          },
          "gpu_count": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Gpu Count",
            "default": 1
          },
          "owner": {
            "type": "string",
            "title": "Owner",
            "default": ""
          },
          "purpose": {
            "type": "string",
            "title": "Purpose",
            "default": "inference"
          },
          "preferred_indices": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Preferred Indices"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "title": "GpuAllocationRequest"
      },
      "GpuAllocationResult": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "allocation_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allocation Ids"
          },
          "gpu_indices": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Gpu Indices"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "GpuAllocationResult"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HeartbeatRequest": {
        "properties": {
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "node_id"
        ],
        "title": "HeartbeatRequest"
      },
...
      "MemoryViewResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id",
            "default": ""
          },
          "conversation": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Conversation"
          },
          "project_memory": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Project Memory"
          },
          "shared": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Shared"
          },
          "knowledge": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Knowledge"
          }
        },
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "MemoryViewResponse"
      },
...
      "PromptViewResponse": {
        "properties": {
          "prompt_id": {
            "type": "string",
            "title": "Prompt Id"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": ""
          },
          "content": {
            "type": "string",
            "title": "Content",
            "default": ""
          },
          "variables": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Variables"
          },
          "overridden": {
            "type": "boolean",
            "title": "Overridden",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "prompt_id",
          "version"
        ],
        "title": "PromptViewResponse"
      },
...
      "ProviderKind": {
        "type": "string",
        "enum": [
          "vllm",
          "sglang"
        ],
        "title": "ProviderKind"
      },
...
      "ProviderRegisterRequest": {
        "properties": {
          "node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Id"
          },
          "kind": {
            "$ref": "#/components/schemas/ProviderKind"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "default": "EMPTY"
          },
          "weight": {
            "type": "integer",
            "title": "Weight",
            "default": 100
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "base_url"
        ],
        "title": "ProviderRegisterRequest"
      },
...
      "RerunStepRequest": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "node_id": {
            "type": "string",
            "title": "Node Id"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "node_id"
        ],
        "title": "RerunStepRequest"
      },
...
      "TraceViewResponse": {
        "properties": {
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "default": ""
          },
          "duration_sec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Sec"
          },
          "timeline": {
            "type": "string",
            "title": "Timeline",
            "default": ""
          },
          "events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Events"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          }
        },
        "type": "object",
        "required": [
          "trace_id"
        ],
        "title": "TraceViewResponse"
      },
      "UpdatePromptRequest": {
        "properties": {
          "prompt_id": {
            "type": "string",
            "title": "Prompt Id"
          },
          "version": {
            "type": "string",
            "title": "Version",
            "default": "1.0.0"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "prompt_id",
          "content"
        ],
        "title": "UpdatePromptRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkflowViewResponse": {
        "properties": {
          "workflow_id": {
            "type": "string",
            "title": "Workflow Id"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "nodes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Edges"
          },
          "node_results": {
            "additionalProperties": true,
            "type": "object",
            "title": "Node Results"
          },
          "pending_node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Node Id"
          }
        },
        "type": "object",
        "required": [
          "workflow_id"
        ],
        "title": "WorkflowViewResponse"
      }
    }
  },
...
}